# Optimize doResize method with binary search 