@JustSNguyen 2024-11-07
I have updated the PR to also support this for the mobile version @gaearon.