@auroursa 2024-11-12
Excellent! Uploading photos taken with an iPhone（obviously compression is needed） clearly shows that the new algorithm improves image quality noticeably, with a shorter processing time as well.