@gaearon 2024-11-25
This is changing `lib/media/manip` but I'm seeing some code go through a similar path in `src/state/gallery`. Would you mind checking how it works today? It seems like maybe it's forked between the image picker vs inline camera, but I'm not sure.